Allow options to be registered passively. (#9061)
A "passive" option is one that cannot (normally) be referenced from
code, does not appear in help output, or specified via flag.
But if specified in config, it does not cause config validation to fail.

This allows us to conditionally register options (e.g., in v1 vs. v2)
without having to edit config files when the condition changes in
order to avoid validation errors.

Also applies this to the previously conditionally-registered options.