[IntelliJ] Add `javac_options` and `extra_jvm_options` to the output of `export-dep-as-jar` (#9053)
### Problem

In addition to scalac options (implemented in #9036), we also need javac options and jvm options in order to faithfully reproduce compilation of source code in outside tools.

### Solution

Added javac_options to the output of export-dep-as-jar, in the same way as scalac options in #9036.
Added extra_jvm_options to the output of export-dep-as-jar, by reading the payload field of targets. Defaults to [] if not present.
Modified integration test suite to allow to test this.
Add tests for both of these usecases.
This is now ready for review. Commit cca7358 (Split zinc argfile by newline instead of space) is #9065, and this PR depends on that one. Other than that, every commit should be independently reviewable.