Fix deprecation warning for `--transitive` for goals that ignore the option (#9076)
In https://github.com/pantsbuild/pants/pull/8955, we deprecated the `--transitive` option for 11 console tasks where the option did not do anything. 

But, a bug in the implementation meant that we would print a deprecation warning unless the user explicitly used `--no-transitive` or `--transitive`, which is deprecated as well, so it was impossible to run these tasks without some deprecation warning.

Also, the tasks would stop working properly once the `--transitive` option is removed because `ConsoleTask` was extending `HasTransitiveOptionsMixin`, which calls `self.get_options().transitive`. This would throw an exception if there was no option `--transitive` defined.