Warn when `fix_deprecated_globs_usage.py` encounters variables (#9080)
Twitter has a couple examples of BUILD files using variables, rather than raw strings, for globs. The script will currently discard those variables, rather than warning that they can't be safely updated.

(The variables cannot be safely updated because we don't know where they are defined and it would be very complex to try to backtrack that information, especially to update exclude variables. Instead, we warn and accept that the user will need to manually fix, which is not very difficult to do.)

End users will get this update because the `curl` link always points to master.