Move more python utils to core (#9082)
The upcoming pex 2.x upgrade needs various utils that currently live in the python backend. We'd rather not have code dependencies from core to a backend, so we move the relevant code into the pants.python package.

This acknowledges the distinction between "things pants needs to know about python because it's written in python" vs. "things pants needs in order to build end-user python code".