Add `migrate_to_toml_config.py` script to automatically update INI config files to TOML (#9054)
This script removes 80% of the tedium of migrating from INI config files to TOML config files. 

It does not attempt to handle the remaining 20%:
- multiline options
- comments
- nested lists and nested dictionaries

Users will be able to manually fix the remaining issues through either the validation from their editor or the site https://www.toml-lint.com.

--

Similar to `fix_deprecated_globs_usage.py`, this is not officially shipped with the `pantsbuild.pants` wheel, but we keep it in this repository for version control, tests, and to have a stable URL that we can instruct users to `curl`.