Handle Ctrl-C in the V2 UI (#9043)
## Problem

The v2 console UI sets the terminal to raw mode, causing it to no longer detect Ctrl-C as a sign that the user is trying to interrupt the process.

## Solution

This commit re-architects the render() method on EngineDisplay to poll for keyboard events every time it runs, and return a sigil value if it detects the key combination Ctrl-C. The scheduler's execute method passes up this potential error value to the scheduler_execute FFI function, which in turn signals that a Ctrl-C has been pressed by returning a null pointer to Python. In Python, the code checks for a null pointer and raises a Python KeyboardInterrupt if one is found.

## Result

Hitting Ctrl-C should now have the same effect as it did in V1, i.e. raising a Python KeyboardInterrupt and quitting the program after cleanup.