Allow specifying paths to lookup for IntelliJ location on MAC
Since this commit, pants on MACs looks up for IntelliJ in two locations
      '/Applications/IntelliJ IDEA CE.app'
      '/Applications/IntelliJ IDEA.app'

If this is not sufficient, a new location can be added via
`--possible-paths` options.

The IntelliJ on MAC is launched with `open -a <path_to_intellij>`
command instead of just running `idea` script, because we don't want
to block shell when IntelliJ is spawned.