Inline `inject_init.py` into `prepare_chrooted_python_sources.py` (#9115)
We were only using `inject_init.py` for the purpose of `prepare_chrooted_python_sources.py`. Inlining this simplifies things. If we ever end up needing the init logic to be used in another context, we can, of course, generalize it back out.

More importantly, this stops using a hacky `ExecuteProcessRequest` to create the missing `__init__.py` files to instead use engine intrinsics. This improves readability and performance.