Add `find_target_source_files` rule to generalize precise file arguments (#9127)
Soon, we will be adding precise file arguments for `fmt2` and `lint2` (meaning we only run over the specified files, rather than the entire target).

When adding this for `test`, we intentionally did not generalize the logic to figure out which specific files should be run against because we were still discovering the best design. Now, we must generalize as prework for `fmt2` and `lint2`.