Remove pytest passthrough args deprecation. (#9126)
### Problem

`pytest` passthrough args were deprecated, but there is a strong possibility (see #9075) that they can be preserved for unambiguous cases in v2.

### Solution

Remove the `pytest` passthrough args deprecation.

Note that this intentionally leaves the `mypy` and `isort` deprecations: the reason for this is that use of passthrough args for `lint` tasks would be effectively guaranteed to be ambiguous given how many there are. Additionally, I expect that we will want the flexibility to begin (optionally) running `mypy` as part of the `test` goal.