Move comma-separated display args to be a field
It appears the introduction of `dataclasses.asdict` in https://github.com/pantsbuild/pants/pull/8853/files\#diff-7c4477315e6112956eaab1f890182117R58 meant that the `comma_separated_display_args` no longer rendered as expected in the template. This moves that into the constructor so we dump the appropriate fields to be consumed by the `single_option.html.mustache` template.