No-op when V2 goals have no implementation (#9142)
### Problem

We want to be able to turn on `--v2` by default to make it easier for users to use V2. This is especially important once we rename `fmt2` with `fmt`. Once that happens, a user would need to use `--v2` to get any of the V2 linters working.

Currently, we can't turn on `--v2` by default because several goals will error when no `backend_packages2` are registered, e.g.

```bash
▶ ./pants --v1 --v2 --pants-config-files='[]' test tests/python/pants_test/util:strutil

ERROR: Not a registered union type: <class 'pants.util.meta.TestTarget'>
```

### Solution

Check in `LegacyGraphSession.run_goal_rules()` if the goal is implemented or not. If not implemented, no-op.

This allows us to get the functionality we want without needing to make any changes to the actual  `@goal_rule`s.


### Result

`./pants --v1 --v2 --pants-config-files='[]' test tests/python/pants_test/util:strutil` will run the V1 implementation and do nothing with the V2 implementation. This means it will now be safe for us to turn on `--v2` by default.

Once we remove V1, we may want to change the behavior to warn to the user that no backends are registered. (Possibly, we should detect if they have `--no-v1` enabled and warn if so.)