Cleanup tests handling of `download_pex_bin.py` (#9143)
Recently, we added support for configuring the version of V2 binary tools like Pex in https://github.com/pantsbuild/pants/pull/8859.

One of the tests was failing because the `download-pex-bin` option scope was not registered. This was because `pants.backend.awslambda.python` used to have an implicit dependency on `pants.backend.python` to work properly; without `pants.backend.python` registered, the AWSLambda backend would not work properly because it would not register all the subsystems and rules needed to stand on its own. This fixes it so that `pants.backend.awslambda.python` will work even if `pants.backend.python` is not registered.

This also cleans up the linter tests to stop special-casing registration of `download_pex_bin.py`.