[export-dep-as-jar] Include transitive target dependencies in export-dep-as-jar
### Problem
There is currently no way for a tool consuming the output of export-dep-as-jar to know whether to traverse the transitive dependency graph of a target or not. The tool shouldn't do that when the target has strict_deps enabled.

### Solution
Add a field to the ouput of the target, "source_dependencies_in_classpath". This field is the intersection between "modulizable targets" and "transitive dependencies", respecting strict_deps.

### Result
The output for the target now has that field, called source_dependencies_in_classpath. Tools can use this field instead of targets when figuring out which targets need including in the classpath. No need to traverse targets now.