Rename Platform -> PlatformConstraint (#9154)
### Problem

Python has a pair of enum types `Platform` (with variants "darwin" and "linux"), and `PlatformConstraint` (with variants "none", "darwin" and "linux"). There is also a rust enum  type defined in `process_execution` called `Platform`, but it has three variants like Python `PlatformConstraint` rather than two. This makes it confusing to reason about how the rust and python types correspond.

### Solution

Rename the Rust type to `PlatformConstraint` to indicate that it is analogous to the Python type of that name