Simplify the solution by no longer using self._map
Turns out, this was an over-optimization and likely not even a good optimization. The author explained that they used it for O(1) .index() and __contains__ operations. But, these things are already very fast on tuples and lists. For example, there was only a 0.0001 slow down between the two implementations for calling `__contains__` on an OrderedSet with 10k random numbers.

Plus, we don't expect to need to call .index() frequently. It's better to reduce the space complexity and slightly increase the speed of the constructor (removing another entire pass over the deduplicated elements).