Fix help printing for v2 goals (#9158)
### Problem

When a user runs `--help` on a v2 goal, pants will current provide help output as if it is a v1 goal - that is, provide help information for the goal and related subystems and v1 tasks. Most of this information is irrelevant in a v2 context and should not be displayed.

### Solution

This commit modifies the help printer to only print help information relevant to v2 if v2 is enabled and v1 is not enabled.