Set up infrastructure for generating coverage reports in V2 `./pants test` (#9170)
### Problem

in #8978 I added coverage support for pytest in V2 but discovered it induced a dep from the generic test runner to the python backend. This adds a generic CoverageReport to test.py and lays the groundwork for finishing the pytest coverage work without inducing the unwanted backend dependency.

Huge thanks to Stu and Eric for helping me figure this bit out! You guys are the best!