Don't require V2 test implementations to implement coverage (#9173)
https://github.com/pantsbuild/pants/pull/9170 required `CoverageDataBatch` to be implemented for the V2 `test` goal to work. Otherwise, it would gracefully no-op.

Because we don't yet have `pytest-cov` implemented in V2, this meant that `./v2 test` stopped working.

Generally, we probably don't want to require that test implementations have coverage support to work. Coverage support is an added bonus and it's meaningful to have a test runner even if it doesn't have coverage.