Refactor `strip_source_roots` to use `TargetAdaptor` rather than `HydratedTarget` (#9178)
### Problem

The rule `strip_source_roots_from_target` currently expects a `HydratedTarget`, but it doesn't actually need that much information/hydration. All it needs is the underlying `TargetAdaptor`.

Requiring `HydratedTarget` makes the API less ergonomic than possible as it requires call-sites that have a `TargetAdaptor` to first `await Get[HydratedTarget](Address, adaptor.address)` (which also has a performance hit).

This is getting in the way of generalizing the `determine_source_files` rules.

### Solution

Add `StripTargetRequest` to complement `StripSnapshotRequest`.