One more round of simplication: only allow file paths
This allows us to support multiple lockfiles for no extra work! At least so long as constraints are only allowed as a global option, we don't anticipate much use for raw constraint strings. With target-level constraints, it would be more helpful, but it's not yet clear what target-level constraints will look like.