Fix plugin resolver cache compatibility.
The cache between Pants on Pex 1.x and Pants on Pex 2.x was incompatible
(in particular the contents of plugins-<hash>.txt file) leading to
errors when upgrading or downgrading pants that forced either clearing
the plugins cache or else modifying its location via setting the global
`plugin_cache_dir` option.

This change fixes the plugin cache to be divorced from the results of
Pex resolves by copying Pex resolved distributions to a location
controlled by the plugin system that is backwards compatible with older
versions of Pants.