Tweak `TargetAdaptor.dependencies` to be a tuple rather than a list (#9184)
### Problem

`TargetAdaptor` and `HydratedTarget` duplicate information - they both have `address` and `dependencies` fields that are almost identical, outside of `TargetAdaptor` using a list rather than a tuple of `Address` instances for the `dependencies` field.

@jsirois explained that there was no particular reason to use lists, rather than tuples, with this `Struct` code from the very first days of the engine. Recently, we've defaulted to always using tuples instead of lists because they are immutable and hashable, but this is not a common idiom in Python so we had used lists in 2015.

### Solution

Rename `@addressable_list` to `@addressable_sequence` (`@addressable_tuple` sounded wrong because that makes it sound like a 2-tuple, 3-tuple, etc). 

Tweak `addressable.py` and `build_files.py` to stop expecting a `MutableSequence` and instead work with either `tuple` or `list`. We don't enforce it gets `list` because we don't want to waste time trying to update all call sites. Then, we cast everything to a `tuple`, unlike previously using a `list`.

Finally, we use a hacky `__init__()` to get MyPy to understand that `TargetAdaptor.dependencies` is a `Tuple[Address, ...]`.

### Result

We'll be able to remove `HydratedTarget.address` and `HydratedTarget.dependencies` as fields because they're both contained within `TargetAdaptor`.

This is sort of prework for the Target API.