Allow mocking unions in `testutil.engine.util.run_rule()` (#9203)
### Problem

The implementation in `run_rule()` requires `type(mock.subject) == mock.subject_type)`. That is, with:

```
await Get[A](B, b)
```

`b` must have exactly the type `B`, regardless of it's a union. This is not how the engine actually works. We allow for either the exact type or for union members.

### Result

The `rules/core` tests now properly mirror their src code rules.