Simplify `fmt2` grouping of formatters per language (#9211)
### Problem

Unlike with `lint`, which can safely run every linter in parallel, formatters must be run sequentially so that they do not overwrite each other. As a performance improvement, we can group formatters by language family: while all Python linters must run sequentially, Scala formatters are not going to touch the same files so those can run sequentially at the same time as Python, and so on.

We previously achieved this grouping by having the concrete class `AggregatedFmtResults` and the union type `FormatTarget`. Each language was expected to give back a single `AggregatedFmtResults` for each target, which stored all the `FmtResult`s for each formatter for that target. `FormatTarget` was a union with members like `PythonTestsAdaptorWithOrigin` and `PythonLibraryAdaptorWithOrigin`. Rules would then convert those target adaptors to an intermediate `PythonFormatTarget`, which was itself a union with members like `IsortTarget`.

This implementation worked but was not as clear as possible.

More importantly, this design makes it challenging to batch multiple targets, as it operates on a per-target basis.

### Solution

Replace `FormatTarget`, `PythonFormatTarget`, and `AggregatedFmtResults` with `LanguageFormatters`, `LanguageFmtResults`, `PythonFormatters`. Rather than `await Get[AggregatedFmtResults](FormatTarget, python_target_adaptor)`, we have `await Get[LanguageFmtResults](LanguageFormatters, PythonFormatters(python_target_adaptor))`. `PythonFormatters` has a rule that runs all union members of `PythonFormatter` and combines them into a result.

Each `LanguageFormatters` subclass defines a `belongs_to_language()` static method to determine whether a target should be run for that language's formatters or not.

### Result

`fmt.py` and `python_formatter.py` are hopefully easier to understand. They are also more closely aligned to `lint.py`.

We can now add the option `--fmt-per-target-caching` to batch multiple targets.