Merge filesystem specs that resolve to the same owning target (#9214)
Closes https://github.com/pantsbuild/pants/issues/9213.

Before:

```
▶ ./v2 list src/python/pants/rules/core/{fmt,lint}.py
src/python/pants/rules/core:core
src/python/pants/rules/core:core
```

After:

```
▶ ./v2 list src/python/pants/rules/core/{fmt,lint}.py
src/python/pants/rules/core:core
```

This is important to ensure that goals behave properly. For example, without this change, `./v2 fmt2 src/python/pants/rules/core/{fmt,lint}.py` will fail because it will result in two different `FmtResult`s for the same target and the engine will fail when trying to merge the two digests, as they share the same files.

--

Here, we add a new type `FilesystemMergedSpec` so that we can preserve the original information, e.g. the original globs, and so that we can know whether this was the originally resolved `FilesystemSpec` or if it was merged from multiple.