Add `--no-fmt-per-target-caching` to run V2 formatters with batched targets (#9218)
Per https://docs.google.com/document/d/1Tdof6jx9aVaOGeIQeI9Gn-8x7nd6LfjnJ0QrbLFBbgc/edit, we found that formatters have substantial startup overhead but have little cost to add an additional file. On a completely cold cache, it is 5-30x slower to run our 3 Python formatters once-per-file rather than in a batch.

https://github.com/pantsbuild/pants/pull/9209 solved this problem for linters by adding `--no-lint-per-target-caching` as an option. Here, we add `--no-fmt-per-target-caching`.

We keep this as an option because it's easy to do that and when remote caching is used it is likely faster to use `--per-target-caching` because there are so many cache hits.