Limit java version to 10 for jax ws gen (#9216)
### Problem
The `JDK/bin/wsimport` is no longer bundled alongside the JDK (since JDK 11), and therefore the whole procedure fails when JAVA_HOME points to a newer JDK. 
I faced this issue when using JDK 11 to run tests of https://github.com/pantsbuild/intellij-pants-plugin. Using JDK 11 there is necessary to [migrate to the newer IntelliJ version](https://github.com/pantsbuild/intellij-pants-plugin/pull/498)

### Solution

The required version of JDK was limited to 10 for JaxWsGen.

### Result

The task will start working when the user is using JDK 11 and there is a JDK older than 11 installed. When there is no such JDK, the procedure will still fail like it failed without this patch, but will give a more descriptive explanation, since it will fail during the JDK resolution and not while trying to run a non-existed executable.