Make GlobalOptions rule provide full options. (#9215)
Previously GlobalOptionsRegistrar was just a bag of classmethods for registration. Actually getting hold of a set of global option values was a little convoluted, especially in the v2 world.

This change makes GlobalOptionsRegistrar a subsystem, so it can actually be instantiated via a subsystem_rule, like any other subsystem. This makes it straightforward to get hold of global option values without special-case handling via OptionsBootstrapper.

As a result, the GlobalOptions rule now provides full options, where previously it only provided bootstrap options, as proven by moving the --colors option from bootstrap to regular options.

Fixes #9212.

Note that this means that in v2 every Optionable is a Subsystem. So when we're fully on v2 we'll be able to greatly simplify the conceptual landscape.