Designate `pants.ini` as legacy, but still supported, in docs (#9194)
We decided to not (yet) deprecate `pants.ini` because it causes too much friction for users at the same time that they're dealing with deprecating `globs()` and `--test-fast`. Further, the code to support `pants.ini` is not very painful to support as it is well-isolated (although, it would allow us to simplify things if we could always assume TOML).

But, we don't want to have to refer to both TOML and INI in our docs and user messages. So, we strike a compromise where we always assume TOML in our docs, but allow our code to work with both. 

To reflect this decision, we update our docs to have a section explaining how INI works and pointing to a script to migrate to TOML. That is the only section in our docs that we should ever mention INI.