Deduplicate `HydratedTarget` and `TargetAdaptor` (#9219)
### Problem

`HydratedTarget` and `TargetAdaptor` currently both duplicate the exact same information in the fields `address` and `dependencies`, i.e. `hydrated_target.address == hydrated_target.adaptor.address` and `hydrated_target.dependencies == hydrated_target.adaptor.dependencies`.

Why keep both these types, then? We need `HydratedTarget` as a simple wrapper around `TargetAdaptor` due to the engine being a type-driven API. Sometimes, we want a generic way to refer to a target, regardless of what the target type is, i.e. `HydratedTarget`. Other times, we do care about the target type, such as a `PythonTestsAdaptor` vs. `PythonLibraryAdaptor`.

So, we need both types, but duplicating the information in both types makes the distinction between the two less clear—we want to emphasize that `HydratedTarget` is nothing more than a wrapper around `TargetAdaptor` for the purpose of having a generic target type for the engine. Duplicating the fields also complicates the implementation, e.g. making it harder to use `HydratedTarget` in tests.

### Solution

Simplify `HydratedTarget` to have only one public* field of `adaptor`. Whenever we need to access the address or dependencies, we can use `hydrated_target.adaptor.address` and `hydrated_target.adaptor.dependencies`.

*`HydratedTarget` gets passed around hundreds of times to the engine when constructing the build graph, so it needs to use `__eq__` hundreds of times to check for cache hits. To improve the speed of this, we override `HydratedTarget.__eq__` to short circuit if the addresses do not match. However, using `adaptor.address` instead of a direct field `._address` results in a slight performance hit when running `build-support/bin/mypy.py` (which runs on every target in Pants). Why? `TargetAdaptor.address` is ~10x slower than `HydratedTarget.address` (0.000_000_1 vs. 0.000_001 seconds). So, we add a private field `HydratedTarget._address` to avoid this performance regression when hashing. Performance stays the same - it takes 4.65 seconds to build the build graph when running `build-support/bin/mypy.py`.

### Result

The difference between `HydratedTarget` and `TargetAdaptor` is hopefully more explicit.

There are no (known) performance regressions for both V1 and V2. It still takes ~4.65 seconds to generate the build graph when running `build-support/bin/mypy.py`.

This is sort of prework for https://github.com/pantsbuild/pants/issues/4535 as it simplifies the status quo.