Don't use ProjectTree in Scheduler (#9202)
The only reason the `Scheduler` requires a `ProjectTree` as input is to pass along the `build_root` and `ignore_patterns` to the native scheduler over the FFI. Since no other functionality of `Scheduler` relies on a `ProjectTree` being passed in, it makes sense to just pass in these two parameters separately, without requiring a `ProjectTree`. This will also make more sense when we support passing an option to the engine to respect `.gitignore` files, via a similar mechanism to how `ignore_patterns` is passed to the engine.