Disable and re-enable async stdin when UI is disabled (#9222)
### Problem

When a user does a pants operation that involves disabling the v2 console UI (such as the `repl` goal or running `test` with the `--debug` flag), they were seeing some bugs with keyboard input. This was caused by the change in https://github.com/pantsbuild/pants/pull/9043 that introduced the termion `async_stdin` abstraction to read console stdout, which spawns a thread to poll from stdin. When we suspend the console UI, this thread was still polling from stdin, and interfering with the ability of whatever subprocess pants was running to also read from stdin, which is why were were bypassing the console UI to begin with.

### Solution

Make the `async_stdin` value on the `EngineDisplay` struct optional, and then set it to `None` on `suspend` and recreate it on `unsuspend`. This results in dropping an instance of `AsyncStdin` when pants suspends the terminal, which kills the background thread reading stdin and removes the source of interference for the subprocess. When we unsuspend, we re-create the async reader just as we previously did when we first initialized the console UI.

### Result

This fixes a currently-existing bug with erratic live keyboard input to a pants subprocess when the v2 UI is activated.