Fix V2 handling of `compatibility` when it's a string rather than list (#9232)
### Problem

We allow `compatibility` to be either a list of strings (ORed) or a single value.

```python
# This works
python_library(
  name="lib",
  compatibility=['CPython==3.7.*']
)

# This should work too, but doesn't
python_library(
  name="lib",
  compatibility='CPython==3.7.*'
)
```

In V2, we assumed it was always a list so a single string value would cause Pants to fail to find an interpreter.

### Solution

Add a typed property `compatibility` to `PythonTargetAdaptor` (and thus its subclasses). In the property, use `ensure_str_list` to handle both single strings and lists of strings.

This design will almost certainly change with the Target API, but it provides a light-weight way to get the functionality we need without committing to any bigger design.

### Result

Interpreter constraints still don't work properly in V2 (https://github.com/pantsbuild/pants/issues/9210), but one of the 3 issues is now resolved.