Stop defaulting to `--transitive` and remove the option when it no-ops (#9236)
Follows up on https://github.com/pantsbuild/pants/pull/8955. This removes the `--transitive` flag for options where it was ignored or where we should always act transitively. This also changes the default for console tasks to `--no-transitive` (except for `filedeps` and `dependencies`, which have a slower deprecation cycle).

This also adds a new deprecation for `--cloc-transitive`, which is the last blocker to replacing V1 `cloc` with the V2 implementation.