Bump process execution timeout from 2m40s to 5 mins (#9230)
### Problem

There was a slow-down in pants somewhere between 1.26.0.dev0 and 1.26.0.dev2 that makes CI consistently timeout; at least in my recent experience, trying to merge a few PRs.

### Solution

Bump the queue buffer time hardcoded in Rust, which gets added to the test’s timeout value. This has the benefit of applying to every remoted test.

### Result

If all goes according to plan, CI passes on this PR and it becomes easier to merge other PRs until the root cause of the slow down is identified and fixed.
At that point in time, reverting this PR is recommended.

Note: I resisted the urge to create a constant for this magic number because this PR is designed to be reverted later. Probably worth a follow-up, though.