fixing scoverage sourcePath to be relative in OSS
Co-authored-by: Ny Saechao <nsaechao@twitter.com>

The problem is also referenced here : https://github.com/scoverage/scalac-scoverage-plugin/pull/275

### Problem

Currently, at compile time, scoverage stores (as a string) the conical sourcePath (path for the source file) for a statement inside the instrument file `scoverage.coverage`. This string is then again required during report generation to get the source file in `ScoverageHtmlWriter` . This breaks in a distributed environment where report generation is happening on a different machine and is unable to locate the source file due to the absolute path.

### Solution

The current fix makes the path relative when storing and turns it back to a conical one (w.r.t cwd) on retrieving. For now, the fix is merged in twitter forked version of scoverage (version 1.0.2-twitter).

### Result

Scoverage can now work in a distributed environment.