Enable zinc to log diagnostics for jvm languages (#9228)
### Problem

We are interested in monitoring the number of warnings and errors per
target for any given jvm language (java and scala).

The first step in this effort is to ensure that the zinc wrapper has the
ability of storing  this data.

### Solution

Add a "diagnostics-out" flag to  zinc with a path to a file. If
provided, warnings and errors will be output in that file as a json
formatted List[PublishDiagnosticsParams].
This is the Language Server Protocol representation of a list of
diagnostics (compilation errors or warnings).
See the schema here:
https://microsoft.github.io/language-server-protocol/specifications/specification-3-14/#textDocument_publishDiagnostics
Liberty is taken to express the URIs to code files in terms of relative
paths to avoid remote cache invalidation.

### Result

This will serve our immediate purpose of aggregating information about
the number of errors and warnings in code built by pants; but could also
have wider applications, such as reporting warnings in case of a cache
hit as the diagnostics file can be dumped in the build cache where it
will be cached as any other compilation artifact.

This PR is currently a no-op from pants' point of view as a new version
of zinc will have to be published and consumed from the python consuming
side before changes take effect.