Mute the path scrub warning when running from sources.
Normally, if PYTHONPATH is set we warn that we're scrubbing it,
so that users don't expect Pants subprocesses to acknowledge
that PYTHONPATH.

However when running from sources we must set PYTHONPATH to the
pants source tree, so this warning is not helpful, and can be
grating.

This change uses another env var to conditionally mute the warning,
and sets that var in the custom "from sources" `./pants` script
in the pants repo.