Open Remove leftover `globs()` code for `bundles` (#9262)
Now that we removed support for `globs`, `rglobs()`, and `zglobs()`, `bundle(fileset=`) will always either be a single string or a list of strings.

Here, we simplify the code for `bundle` and its tests to remove handling for `globs()`. 

Notably, this removes an edge case where users use an eagerly-evaluated `globs()` but also set `rel_path`. This case is correctly handled by the V2 engine code:

https://github.com/pantsbuild/pants/blob/ff5ff5cf4475f4a97bd1cf9df678fa2aaee9e0c3/src/python/pants/engine/legacy/structs.py#L254-L268