Rename `fmt2` to `fmt` and `lint2` to `lint` (#9257)
Closes https://github.com/pantsbuild/pants/issues/8351.

It's safe to make this change, even with us defaulting to `--v2` for users, because we default to no V2 linters being enabled and the V2 goal no-ops in this case. 

It's also safe to have both V1 and V2 linters run at the same time (mixed mode). They won't run at the same time, so there's no risk of formatters conflicting with each other.

In a followup, we can now deprecate V1 isort for the V2 implementation.