Support build file aliases in v2.
This is currently necessary in order to support the
python_requirements() macro.  However this may be
superseded by the v2 target API, so this is more of
a stopgap.

Note that we introduce a new hook, build_file_aliases2,
instead of using the existing build_file_aliases.  This is
so we can minimally introduce just what we need to get
things working in v2-exclusive mode, until the full v2
target API is implemented.

Also fixes some nits in list_targets.py.