Remove stale service-shutdown comment, and log the error stacktrace. (#9272)
## Problem
When pantsd's PailgunService shuts down abnormally, it does not currently log the complete stacktrace for the error that killed it. The comment next to the log message indicates that this is probably because we used to explicitly shutdown the socket server.

## Solution
Since we no longer explicitly shut down the socket server when pantsd is exiting, any exception in this codepath is unexpected, fatal, and worth debugging. Log it.