Lmdb versioning (#9269)
### Problem

In order to implement https://github.com/pantsbuild/pants/pull/9198 , which stores the Platform a remote process executed on in lmdb, we need to change the way that bytes we store to/load from lmdb are (de)serialized to rust data structures. This is fundamentally a database schema change, which we don't currently have a way to do gracefully.

### Solution

This commit introduces a new type `VersionedFingerprint`, which is a simple concatenation of the existing `Fingerprint` type and a one-byte schema version tag. From here on out, if we ever need to change the way we store a type in lmdb, we can bump the schema version (and add an additional byte if we ever need to do this more than 255 times), which will make existing `Fingerprint`s stored in the database no longer map to the same effective keys they did previously. This will make it so that when pants upgrades (or downgrades),  pants won't find a value stored in lmdb under a different data model and try to deserialize it incorrectly.

This commit also renames the currently-existing instantiations of lmdb, which on a one-time basis with this commit will effectively invalidate all existing cache entries.